# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AVIV Clan+ API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface User {
  id: number;
  discordId: string;
  username: string;
  /** @nullable */
  avatar: string | null;
}

export interface ClanSummary {
  id: number;
  name: string;
  /** @nullable */
  imageUrl: string | null;
  memberCount: number;
  isPrivate: boolean;
  alertCount: number;
}

export interface Clan {
  id: number;
  name: string;
  /** @nullable */
  imageUrl: string | null;
  leaderId: number;
  leaderUsername: string;
  isPrivate: boolean;
  /** @nullable */
  discordServerId: string | null;
  /** @nullable */
  discordChannelId: string | null;
  /** @nullable */
  discordChannelName: string | null;
  pingRole: string;
  memberCount: number;
  alertCount: number;
  /** @nullable */
  lastAlertAt: string | null;
  /**
     * The current user role in this clan or null if not a member
     * @nullable
     */
  myRole: string | null;
  isSilenced: boolean;
  hasRaidKey: boolean;
  createdAt: string;
}

export interface ClanInput {
  /**
     * @minLength 1
     * @maxLength 40
     */
  name: string;
  imageUrl?: string;
  isPrivate?: boolean;
}

export interface ClanUpdate {
  /**
     * @minLength 1
     * @maxLength 40
     */
  name?: string;
  imageUrl?: string;
  isPrivate?: boolean;
  raidKey?: string;
  kaosApiKey?: string;
  discordServerId?: string;
  discordChannelId?: string;
  discordChannelName?: string;
  pingRole?: string;
}

export type ClanMemberRole = typeof ClanMemberRole[keyof typeof ClanMemberRole];


export const ClanMemberRole = {
  leader: 'leader',
  member: 'member',
} as const;

export interface ClanMember {
  userId: number;
  clanId: number;
  username: string;
  /** @nullable */
  avatar: string | null;
  role: ClanMemberRole;
  silenced: boolean;
  joinedAt: string;
}

export interface MembershipUpdate {
  silenced?: boolean;
}

export type MyClanEntryRole = typeof MyClanEntryRole[keyof typeof MyClanEntryRole];


export const MyClanEntryRole = {
  leader: 'leader',
  member: 'member',
} as const;

export interface MyClanEntry {
  id: number;
  name: string;
  /** @nullable */
  imageUrl: string | null;
  role: MyClanEntryRole;
  silenced: boolean;
  memberCount: number;
  pendingRequestCount: number;
  /** @nullable */
  lastAlertAt: string | null;
  alertCount: number;
  hasRaidKey: boolean;
  discordConfigured: boolean;
}

export interface AlertEntry {
  id: number;
  clanId: number;
  clanName: string;
  title: string;
  body: string;
  /** @nullable */
  serverId: string | null;
  isTest: boolean;
  createdAt: string;
}

export interface DashboardSummary {
  totalClans: number;
  totalAlerts: number;
  recentAlerts: AlertEntry[];
  pendingRequests: number;
}

export type JoinRequestStatus = typeof JoinRequestStatus[keyof typeof JoinRequestStatus];


export const JoinRequestStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface JoinRequest {
  id: number;
  clanId: number;
  userId: number;
  username: string;
  /** @nullable */
  avatar: string | null;
  status: JoinRequestStatus;
  createdAt: string;
}

export type JoinRequestResolutionAction = typeof JoinRequestResolutionAction[keyof typeof JoinRequestResolutionAction];


export const JoinRequestResolutionAction = {
  approve: 'approve',
  reject: 'reject',
} as const;

export interface JoinRequestResolution {
  action: JoinRequestResolutionAction;
}

export interface InviteTokenInfo {
  token: string;
  url: string;
}

export interface InviteClanPreview {
  clanId: number;
  name: string;
  /** @nullable */
  imageUrl: string | null;
  memberCount: number;
  leaderUsername: string;
}

export interface DiscordChannel {
  id: string;
  name: string;
}

export interface DiscordRole {
  id: string;
  name: string;
  color: number;
}

export interface PushSubscriptionInput {
  endpoint: string;
  p256dh: string;
  auth: string;
}

export interface PushUnsubscribeInput {
  endpoint: string;
}

export interface VapidKeyInfo {
  publicKey: string;
}

export type SearchClansParams = {
/**
 * Name search string
 */
q?: string;
};