# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AVIV Clan+ API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the current authenticated user
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /clans:
    get:
      operationId: searchClans
      tags: [clans]
      summary: Search public clans
      parameters:
        - name: q
          in: query
          description: Name search string
          schema:
            type: string
      responses:
        "200":
          description: Matching public clans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClanSummary"
    post:
      operationId: createClan
      tags: [clans]
      summary: Create a new clan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clan"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clans/{clanId}:
    get:
      operationId: getClan
      tags: [clans]
      summary: Get clan details
      parameters:
        - $ref: "#/components/parameters/ClanId"
      responses:
        "200":
          description: Clan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClan
      tags: [clans]
      summary: Update clan settings (leader only)
      parameters:
        - $ref: "#/components/parameters/ClanId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClanUpdate"
      responses:
        "200":
          description: Updated clan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clan"
        "403":
          description: Not the clan leader
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteClan
      tags: [clans]
      summary: Delete a clan (leader only)
      parameters:
        - $ref: "#/components/parameters/ClanId"
      responses:
        "204":
          description: Deleted
        "403":
          description: Not the clan leader
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Members ───────────────────────────────────────────────────────────────
...
  /clans/{clanId}/discord/channels:
    get:
      operationId: listDiscordChannels
      tags: [discord]
      summary: List text channels in the clan Discord server (requires bot to be in the server)
      parameters:
        - $ref: "#/components/parameters/ClanId"
      responses:
        "200":
          description: Discord channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscordChannel"
        "400":
          description: Bot not in server or no server configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clans/{clanId}/discord/roles:
    get:
      operationId: listDiscordRoles
      tags: [discord]
      summary: List roles in the clan Discord server (requires bot to be in the server)
      parameters:
        - $ref: "#/components/parameters/ClanId"
      responses:
        "200":
          description: Discord roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscordRole"
        "400":
          description: Bot not in server or no server configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Alerts ────────────────────────────────────────────────────────────────
  /clans/{clanId}/alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List recent raid alerts for a clan
      parameters:
        - $ref: "#/components/parameters/ClanId"
      responses:
        "200":
          description: Recent alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertEntry"
  /clans/{clanId}/alerts/test:
    post:
      operationId: sendTestAlert
      tags: [alerts]
      summary: Fire a test raid alert for the clan (leader only)
      parameters:
        - $ref: "#/components/parameters/ClanId"
      responses:
        "200":
          description: Test alert sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertEntry"
        "400":
          description: Discord not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Push subscriptions ────────────────────────────────────────────────────
...
components:
  parameters:
    ClanId:
      name: clanId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        id:
          type: integer
        discordId:
          type: string
        username:
          type: string
        avatar:
          type: ["string", "null"]
      required: [id, discordId, username, avatar]
    ClanSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
        memberCount:
          type: integer
        isPrivate:
          type: boolean
        alertCount:
          type: integer
      required: [id, name, imageUrl, memberCount, isPrivate, alertCount]
    Clan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
        leaderId:
          type: integer
        leaderUsername:
          type: string
        isPrivate:
          type: boolean
        discordServerId:
          type: ["string", "null"]
        discordChannelId:
          type: ["string", "null"]
        discordChannelName:
          type: ["string", "null"]
        pingRole:
          type: string
        memberCount:
          type: integer
        alertCount:
          type: integer
        lastAlertAt:
          type: ["string", "null"]
        myRole:
          type: ["string", "null"]
          description: The current user role in this clan or null if not a member
        isSilenced:
          type: boolean
        hasRaidKey:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - imageUrl
        - leaderId
        - leaderUsername
        - isPrivate
        - discordServerId
        - discordChannelId
        - discordChannelName
        - pingRole
        - memberCount
        - alertCount
        - lastAlertAt
        - myRole
        - isSilenced
        - hasRaidKey
        - createdAt
    ClanInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 40
        imageUrl:
          type: string
        isPrivate:
          type: boolean
      required: [name]
    ClanUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 40
        imageUrl:
          type: string
        isPrivate:
          type: boolean
        raidKey:
          type: string
        kaosApiKey:
          type: string
        discordServerId:
          type: string
        discordChannelId:
          type: string
        discordChannelName:
          type: string
        pingRole:
          type: string
...
    DiscordChannel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required: [id, name]
    DiscordRole:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: integer
      required: [id, name, color]
    AlertEntry:
      type: object
      properties:
        id:
          type: integer
        clanId:
          type: integer
        clanName:
          type: string
        title:
          type: string
        body:
          type: string
        serverId:
          type: ["string", "null"]
        isTest:
          type: boolean
        createdAt:
          type: string
      required: [id, clanId, clanName, title, body, serverId, isTest, createdAt]
...